requirements.txt
langchain
langchain-groq
python-dotenv
langchain-cohere
langchain-anthropic
langchain-huggingface
langchain-google-firestore
langchainhub
wikipedia
langchain-chroma
langchain-pinecone
langchain-community
pymupdf
langchain-huggingface
sentence-transformers